# BE_Desiki/be_desiki/swagger-editor.yaml
openapi: 3.0.0
info:
  title: Account API
  version: 1.0.0
...
paths:
...
  /api/Product/products:
    get:
      tags:
        - ProductController
      summary: Lấy danh sách sản phẩm
      responses:
        '200':
          description: Danh sách sản phẩm
          content:
            application/json:
              schema:
                type: object
                properties:
                  products:
                    type: array
                    items:
                      type: object
                      properties:
                        product:
                          allOf:
                            - $ref: '#/components/schemas/Product'
                            - type: object
                              properties:
                                imageUrl:
                                  type: string
                        category:
                          $ref: '#/components/schemas/Category'
                        shipmentProducts:
                          type: array
                          items:
                            type: object
                            properties:
                              shipment:
                                $ref: '#/components/schemas/Shipment'
                              shipmentProduct:
                                $ref: '#/components/schemas/ShipmentProduct'
                            
                        productSkinTypes:
                          type: array
                          items:
                            $ref: '#/components/schemas/SkinType'
                        productSkinStatuses:
                          type: array
                          items:
                            $ref: '#/components/schemas/SkinStatus'
    post:
      tags:
        - ProductController
      summary: Thêm sản phẩm mới
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                product:
                  type: object
                  properties:
                    categoryId:
                      type: integer
                      nullable: true
                    name:
                      type: string
                    description:
                      type: string
                      nullable: true
                    volume:
                      type: number
                      format: float
                      nullable: true
                    salePrice:
                      type: number
                      format: double
                    gameTicketReward:
                      type: number
                    imageBase64:
                      type: string
                skinTypeIds:
                  type: array
                  items:
                    type: integer
                skinStatusIds:
                  type: array
                  items:
                    type: integer
      responses:
        '200':
          description: Sản phẩm được thêm thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /api/Product/products/{productId}:
    get:
      tags:
        - ProductController
      summary: Lấy chi tiết sản phẩm
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Chi tiết sản phẩm
          content:
            application/json:
              schema:
                type: object
                properties:
                  product:
                    allOf:
                      - $ref: '#/components/schemas/Product'
                      - type: object
                        properties:
                          imageUrl:
                            type: string
                  category:
                          $ref: '#/components/schemas/Category'
                  shipmentProducts:
                    type: array
                    items:
                      type: object
                      properties:
                        shipment:
                          $ref: '#/components/schemas/Shipment'
                        shipmentProduct:
                          $ref: '#/components/schemas/ShipmentProduct'                    
                  productSkinTypes:
                    type: array
                    items:
                      $ref: '#/components/schemas/SkinType'
                  productSkinStatuses:
                    type: array
                    items:
                      $ref: '#/components/schemas/SkinStatus'
    put:
      tags:
        - ProductController
      summary: Cập nhật sản phẩm
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                product:
                  type: object
                  properties:
                    categoryId:
                      type: integer
                      nullable: true
                    name:
                      type: string
                    description:
                      type: string
                      nullable: true
                    volume:
                      type: number
                      format: float
                      nullable: true
                    salePrice:
                      type: number
                      format: double
                    gameTicketReward:
                      type: number
                    imageBase64:
                      type: string
                skinTypeIds:
                  type: array
                  items:
                    type: integer
                skinStatusIds:
                  type: array
                  items:
                    type: integer
      responses:
        '200':
          description: Cập nhật sản phẩm thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /api/Product/products/{productId}/deactivate/{isDeactivate}:
    put:
      tags:
        - ProductController
      summary: Vô hiệu/kích hoạt sản phẩm
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
        - name: isDeactivate
          in: path
          required: true
          schema:
            type: boolean
      responses:
        '200':
          description: Đổi trạng thái sản phẩm thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /api/Product/shipments:
    get:
      tags:
        - ProductController
      summary: Lấy danh sách lô hàng
      responses:
        '200':
          description: Danh sách lô hàng
          content:
            application/json:
              schema:
                type: object
                properties:
                  shipments:
                    type: array
                    items:
                      type: object
                      properties:
                        shipment:
                          $ref: '#/components/schemas/Shipment'
                        shipmentProducts:
                          type: array
                          items:
                            type: object
                            properties:
                              shipmentProduct:
                                $ref: '#/components/schemas/ShipmentProduct' 
                              product:
                                allOf:
                                  - $ref: '#/components/schemas/Product'
                                  - type: object
                                    properties:
                                      imageUrl:
                                        type: string
                              category:
                                $ref: '#/components/schemas/Category'

    post:
      tags:
        - ProductController
      summary: Thêm lô hàng
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                shipment:
                  type: object
                  properties:
                    _id:
                      type: string
                    shipmentDate:
                      type: string
                      format: date

      responses:
        '200':
          description: Lô hàng được thêm thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /api/Product/shipments/{shipmentId}:
    get:
      tags:
        - ProductController
      summary: Lấy chi tiết lô hàng
      parameters:
        - name: shipmentId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Chi tiết lô hàng
          content:
            application/json:
              schema:
                type: object
                properties:
                  shipment:
                    $ref: '#/components/schemas/Shipment'
                  shipmentProducts:
                    type: array
                    items:
                      type: object
                      properties:
                        shipmentProduct:
                          $ref: '#/components/schemas/ShipmentProduct' 
                        product:
                          allOf:
                            - $ref: '#/components/schemas/Product'
                            - type: object
                              properties:
                                imageUrl:
                                  type: string
                        category:
                          $ref: '#/components/schemas/Category'
    put:
      tags:
        - ProductController
      summary: Cập nhật lô hàng
      parameters:
        - name: shipmentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                shipment:
                  type: object
                  properties:
                    shipmentDate:
                      type: string
                      format: date
      responses:
        '200':
          description: Chi tiết lô hàng
          content:
            application/json:
              schema:
                type: object
                properties:
                  shipment:
                    $ref: '#/components/schemas/Shipment'
                  shipmentProducts:
                    type: array
                    items:
                      type: object
                      properties:
                        shipmentProduct:
                          $ref: '#/components/schemas/ShipmentProduct' 
                        product:
                          allOf:
                            - $ref: '#/components/schemas/Product'
                            - type: object
                              properties:
                                imageUrl:
                                  type: string
                        category:
                          $ref: '#/components/schemas/Category'
      
    delete:
      tags:
        - ProductController
      summary: Xóa lô hàng
      parameters:
        - name: shipmentId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Xóa lô hàng thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /api/Product/shipmentProducts/{shipmentProductId}:
    get:
      tags:
        - ProductController
      summary: Lấy chi tiết shipmentProduct
      parameters:
        - name: shipmentProductId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Chi tiết shipmentProduct
          content:
            application/json:
              schema:
                type: object
                properties:
                  shipmentProduct:
                    $ref: '#/components/schemas/ShipmentProduct' 
                  shipment:
                          $ref: '#/components/schemas/Shipment'
                  product:
                    allOf:
                      - $ref: '#/components/schemas/Product'
                      - type: object
                        properties:
                          imageUrl:
                            type: string
    put:
      tags:
        - ProductController
      summary: Cập nhật shipmentProduct
      parameters:
        - name: shipmentProductId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                shipmentProduct:
                  type: object
                  properties:
                    importQuantity:
                      type: integer
                    manufacturingDate:
                      type: string
                      format: date
                      nullable: true
                    expiryDate:
                      type: string
                      format: date
                      nullable: true
                    buyPrice:
                      type: number
                      format: double
                      nullable: true
      responses:
        '200':
          description: Cập nhật shipmentProduct thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /api/Product/shipmentProducts:
    post:
      tags:
        - ProductController
      summary: Đăng ký shipmentProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                shipmentProduct:
                  type: object
                  properties:
                    productId:
                      type: string
                      format: objectId
                    shipmentId:
                      type: string
                    importQuantity:
                      type: integer
                    manufacturingDate:
                      type: string
                      format: date
                      nullable: true
                    expiryDate:
                      type: string
                      format: date
                      nullable: true
                    buyPrice:
                      type: number
                      format: double
                      nullable: true
      responses:
        '200':
          description: Đăng ký shipmentProduct thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /api/Product/shipmentProducts/{shipmentProductId}/deactivate/{isDeactivate}:
    put:
      tags:
        - ProductController
      summary: Vô hiệu/kích hoạt shipmentProduct
      parameters:
        - name: shipmentProductId
          in: path
          required: true
          schema:
            type: string
        - name: isDeactivate
          in: path
          required: true
          schema:
            type: boolean
      responses:
        '200':
          description: Đổi trạng thái shipmentProduct thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
...
components:
  schemas:
...
    Shipment:
      type: object
      properties:
        _id:
          type: string
        shipmentDate:
          type: string
          format: date
        isDeleted:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ShipmentProduct:
      type: object
      properties:
        _id:
          type: string
          format: objectId
        productId:
          type: string
          format: objectId
        shipmentId:
          type: string
        importQuantity:
          type: integer
        saleQuantity:
          type: integer
        manufacturingDate:
          type: string
          format: date
          nullable: true
        expiryDate:
          type: string
          format: date
          nullable: true
        buyPrice:
          type: number
          format: double
          nullable: true
        isDeactivated:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Product:
      type: object
      properties:
        _id:
          type: string
          format: objectId
        categoryId:
          type: integer
          nullable: true
        name:
          type: string
        description:
          type: string
          nullable: true
        volume:
          type: number
          format: float
          nullable: true
        salePrice:
          type: number
          format: double
        gameTicketReward:
          type: number
        isDeactivated:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Category:
      type: object
      properties:
        _id:
          type: integer
        name:
          type: string
    SkinType:
      type: object
      properties:
        _id:
          type: integer
        name:
          type: string
    SkinStatus:
      type: object
      properties:
        _id:
          type: integer
        name:
          type: string
...